[REFACTOR] 복수형 클래스 네이밍 삭제로 인한 수정
