+++ src/main/java/kr/hhplus/be/server/application/order/OrderFacade.java
@@ -1,10 +1,9 @@
 package kr.hhplus.be.server.application.order;
 
-import kr.hhplus.be.server.domain.common.exception.DomainExceptions;
+import kr.hhplus.be.server.domain.common.exception.DomainException;
 import kr.hhplus.be.server.domain.order.Order;
 import kr.hhplus.be.server.domain.user.User;
 import kr.hhplus.be.server.domain.user.UserRepository;
-import kr.hhplus.be.server.infrastructure.user.UserJpaRepository;
 import kr.hhplus.be.server.interfaces.api.order.OrderProductRequest;
 import kr.hhplus.be.server.interfaces.api.order.OrderResponse;
 import lombok.RequiredArgsConstructor;
@@ -12,6 +11,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+
+import static kr.hhplus.be.server.domain.common.exception.DomainException.*;
 
 @Service
 @RequiredArgsConstructor
@@ -23,7 +24,7 @@ public class OrderFacade {
     public OrderResponse createOrder(Long userId, List<OrderProductRequest> orderProductRequests) {
         // 사용자 조회: 존재하지 않으면 도메인 예외 발생
         User user = userJpaRepository.findById(userId)
-                .orElseThrow(() -> new DomainExceptions.EntityNotFoundException("User not found with id: " + userId));
+                .orElseThrow(() -> new EntityNotFoundException("User not found with id: " + userId));
 
         String orderNumber = generateOrderNumber();
         Order order = orderService.placeOrder(user, orderNumber, orderProductRequests);
