+++ src/main/java/kr/hhplus/be/server/domain/point/UserPoint.java
@@ -3,7 +3,7 @@ package kr.hhplus.be.server.domain.point;
 import jakarta.persistence.*;
 import kr.hhplus.be.server.domain.common.BaseEntity;
 import kr.hhplus.be.server.domain.common.Money;
-import kr.hhplus.be.server.domain.common.exception.DomainExceptions;
+import kr.hhplus.be.server.domain.common.exception.DomainException;
 import kr.hhplus.be.server.domain.user.User;
 import kr.hhplus.be.server.infrastructure.point.UserPointRequest;
 import lombok.Getter;
@@ -11,6 +11,8 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.math.BigDecimal;
+
+import static kr.hhplus.be.server.domain.common.exception.DomainException.*;
 
 @Entity
 @Table(name = "user_points")
@@ -43,14 +45,14 @@ public class UserPoint extends BaseEntity {
 
     public void chargePoints(Money amount) {
         if (amount.amount().compareTo(BigDecimal.ZERO) <= 0) {
-            throw new DomainExceptions.InvalidStateException("충전 포인트는 0 이상이어야 합니다.");
+            throw new InvalidStateException("충전 포인트는 0 이상이어야 합니다.");
         }
         this.pointBalance = new Money(this.pointBalance.amount().add(amount.amount()));
     }
 
     public void usePoints(Money amount) {
         if (pointBalance.subtract(amount).compareTo(Money.ZERO) < 0) {
-            throw new DomainExceptions.InvalidStateException("사용 포인트가 부족합니다.");
+            throw new InvalidStateException("사용 포인트가 부족합니다.");
         }
         pointBalance = pointBalance.subtract(amount);
     }
